+++ client/package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "dev": "parcel ./index.html",
+    "dev": "parcel ./index.html --log-level 5 --open",
     "build": "parcel build ./index.html",
     "test": "jest",
     "lint": "eslint --cache --format codeframe --ext mjs,jsx,js src test"
@@ -14,7 +14,7 @@
     "@babel/polyfill": "^7.8.7",
     "@babel/preset-env": "^7.9.0",
     "@babel/preset-react": "^7.9.4",
-    "@date-io/date-fns": "^2.5.0",
+    "@date-io/date-fns": "1.x",
     "@material-ui/core": "^4.9.7",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/pickers": "^3.2.10",
@@ -25,7 +25,7 @@
     "axios": "^0.19.2",
     "babel-loader": "^8.1.0",
     "babel-preset-react": "^6.24.1",
-    "date-fns": "^2.0.0",
+    "date-fns": "^2.11.1",
     "html-webpack-plugin": "^4.0.3",
     "i18next": "^19.3.4",
     "i18next-browser-languagedetector": "^4.0.2",
@@ -54,6 +54,7 @@
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.19.0",
+    "eslint-watch": "^6.0.1",
     "jest": "^25",
     "parcel-bundler": "^1.12.4",
     "prettier": "^2.0.1",
