+++ client/src/components/index.jsx
@@ -2,8 +2,8 @@ export { default as Footer } from './Footer';
 export { default as Header } from './Header';
 export { default as SpecialButton } from './Buttons/Special';
 export { default as HandleActionButton } from './Buttons/HandleAction';
-export { default as DatePicker } from './DatePicker';
 export { default as Step1 } from './Steps/step1';
 export { default as Step6 } from './Steps/step6';
 export { default as CountrySelect } from './CountrySelect';
 export { default as RadioGroup } from './RadioGroup';
+export { default as DatePicker } from './DatePicker';
