+++ client/src/views/index.jsx
@@ -1,6 +1,3 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as Home } from './Home';
 export { default as Questionnary } from './Questionnary';
 export { default as Question } from './Question';
-export { default as TranslationTest } from './TranslationTest';
-export { default as Foobar } from './Foobar';
